Build changes for compatibility with bazel rules_rust
On Cuttlefish we are looking at adding a source dependency on crosvm to
our bazel build for our project hosted in GitHub. The objective is to
use a crosvm built from upstream sources rather than the one which comes
with the Android build, as the now-infrequent AOSP builds will end up
with stale versions of crosvm.

We have a prototype implementation (
https://github.com/google/android-cuttlefish/pull/1044 ) that pulls in
this dependency, but it has some constraints from using `rules_rust`,
which provides Cargo functionality in the context of bazel.

`jail` gets a new environment variable to pass a
`compile_seccomp_policy.py` script, as bazel operates in terms of files
rather than directories.

`= '*'` versions in dependencies are replaced with explicit paths. I'm
don't know if this is viable upstream, the `ebuild` comments imply these
dependencies are substituted with something else in the ChromeOS build
system.

Bug: b/402274999
Change-Id: Ie20d3468617ebbb5692f1a360208cd3c3459b72b